Power ups and some changes
Added - Power ups which includes  Health, Speed, and Invincible.
Health increases the health for character
Speed increases the speed for character
Invincible makes the character immune to the rain

Item check script for the enum

Added in characterStats a check if the player is invincible

Added in rainDetectnDmg an if for when the player is Invincible and when the player is not Invincible.

further explanation inside each scripts